+++ docs/RELEASE-NOTES.rst
@@ -23,7 +23,7 @@ Bug Fixes
 * Fix to remove old ingress monitor when type gets modified
 * Fix to send AS3 declaration for the recreated domain after IPAM controller restart
 * :issues:`2581` IPAM to provide the same IP for different TS
-
+* :issues: `2609` TargetPort support for string with NPL
 
 2.10.1
 -------------